MVP complete. Formatted with Prettier.
