+++ config/secrets.js
@@ -1,3 +1,5 @@
+// DON'T NEED THIS FILE BECAUSE IT'S ON authenticate.js (jwtKey)
+
 module.exports = {
     jwtSecret: process.env.JWT_SECRET || "i like bananas"
   };
